pyproject.toml
[project]
name = "lms-alternative"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "ASTRALLIBERTAD", email = "astrallibertad@gmail.com" }
]
dependencies = [
  "flet == 0.28.3",
  "google-auth",
  "google-auth-oauthlib",
  "google-api-python-client",
]

[tool.flet]
org = "com.pld1"
product = "LMS Alternative"
company = "PLD Group 1"
copyright = "Copyright (C) 2025 by PLD Group 1"
compile.app = true 
compile.packages = true 
compile.cleanup = true 

[tool.flet.app]
path = "src"

[tool.flet.app.startup_screen]
show = true
message = "Starting up the app…"

[tool.flet.app.android]
"android.permission.INTERNET" = true
"android.permission.READ_EXTERNAL_STORAGE" = true
"android.permission.WRITE_EXTERNAL_STORAGE" = true


[build-system]
requires = ["flet"]
build-backend = "flet"